fix(fin-040): refactor profile settings components and improve layout and theme handling
