+++ src/app/(profile)/layout.tsx
@@ -14,7 +14,7 @@ export default function UserLayoutPage({ children }: ChildrenComponentProps) {
   return (
     <AuthGuard>
       <AuthorizedUserProvider>
-        <div className={cn('size-full flex', isMobile && 'h-dvh overflow-hidden flex-col')}>
+        <div className={cn('size-full flex profile-layout', isMobile && 'h-dvh overflow-hidden flex-col')}>
           {!isMobile && <ProfileSidebar />}
 
           <div className="flex-1 min-h-0">{children}</div>
