+++ src/app/(profile)/profile/settings/page.tsx
@@ -1,10 +1,5 @@
 import { ProfileSettingsScreen } from '@frontend/features/setting-page/profile-settings-screen';
-import { ProfileSettingsProvider } from '@frontend/features/setting-page/profile-settings.hook';
 
-export default function SettingPage() {
-  return (
-    <ProfileSettingsProvider>
-      <ProfileSettingsScreen />
-    </ProfileSettingsProvider>
-  );
+export default function SettingsPage() {
+  return <ProfileSettingsScreen />;
 }
